requirements.txt
astroid==2.0.4
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-otp==0.5.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
mysqlclient==1.3.13
pep8==1.7.1
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
